// src/api/generated/model/createIngredientBody.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Bistro App Backend API
 * OpenAPI contract for generated frontend API clients.
 * OpenAPI spec version: 0.1.0
 */
import type { CreateIngredientBodyBaseUnit } from './createIngredientBodyBaseUnit';

export interface CreateIngredientBody {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  barcode?: string | null;
  baseUnit?: CreateIngredientBodyBaseUnit;
  /** @minimum 0 */
  kcalPer100: number;
  /** @minimum 0 */
  proteinPer100: number;
  /** @minimum 0 */
  carbsPer100: number;
  /** @minimum 0 */
  fatPer100: number;
  /**
     * @minimum 0
     * @nullable
     */
  servingGrams?: number | null;
  /** @nullable */
  servingLabel?: string | null;
}